% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/original_data_test.R
\name{print.original_test}
\alias{print.original_test}
\title{Print an \code{original_test} Object}
\usage{
\method{print}{original_test}(x, ...)
}
\arguments{
\item{x}{An object of class \code{original_test}.}

\item{...}{Further arguments, currently ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints a concise summary of a classical test result.
}
