% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_measures.R
\name{print.ps_utility}
\alias{print.ps_utility}
\title{Print Method for \code{ps_utility} Objects}
\usage{
\method{print}{ps_utility}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ps_utility}.}

\item{...}{Further arguments (ignored).}
}
\description{
Print Method for \code{ps_utility} Objects
}
