% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_test_class.R
\name{is.ps_test}
\alias{is.ps_test}
\title{Test if an Object is of Class \code{ps_test}}
\usage{
is.ps_test(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Logical.
}
\description{
Test if an Object is of Class \code{ps_test}
}
